requirements.txt
streamlit
numpy
pillow
pytesseract
opencv-python-headless
transformers
torch
sentencepiece
regex
scikit-learn
